HGD/requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.4.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.12.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Optional: for visualization
matplotlib>=3.7.0

